Changed the `api` routes
