+++ app/admin/[organization]/components/EditEventButton.tsx
@@ -1,15 +1,17 @@
 'use client'
+
 import React, { useContext } from 'react'
-import CreateEventForm from './CreateEventForm' // Assuming you have this component
+import EditEventForm from './EditEventForm'
 import { ModalContext } from '@/components/context/ModalContext'
 import { IEvent } from '@/server/model/event'
+
 const EditOrganizationButton = ({ event }: { event: IEvent }) => {
   const { openModal } = useContext(ModalContext)
   return (
     <div>
       <button
         className="p-2 ml-2 bg-yellow-500 text-white rounded"
-        onClick={() => openModal(<CreateEventForm organizationId={event.organizationId} event={event} />)}>
+        onClick={() => openModal(<EditEventForm organizationId={event.organizationId} event={event} />)}>
         Edit
       </button>
     </div>
