+++ app/admin/[organization]/components/EditEventForm.tsx
@@ -10,7 +10,7 @@ interface EventFormProps {
   event?: IEvent
 }
 
-const CreateEventForm: React.FC<EventFormProps> = ({ organizationId, event }) => {
+const EditEventForm: React.FC<EventFormProps> = ({ organizationId, event }) => {
   const [formData, setFormData] = useState<Omit<IEvent, 'id'>>({
     organizationId: organizationId,
     name: '',
@@ -162,4 +162,4 @@ const CreateEventForm: React.FC<EventFormProps> = ({ organizationId, event }) =>
   )
 }
 
-export default CreateEventForm
+export default EditEventForm
